Replace existing screen with new implementation
Fixing some tests/making them compile in the process.
